+++ src/directive.ts
@@ -2,6 +2,7 @@ import {
   annotate, Declarations, defineMetadata, getAttributeName, getMetadata, isAttributeSelector, kebabToCamel,
   metadataKeys
 } from './utils';
+import { IHostListeners } from './hostListener';
 
 export interface DirectiveOptionsDecorated extends ng.IDirective {
   selector: string;
@@ -14,6 +15,7 @@ export interface DirectiveControllerConstructor {
 export interface DirectiveController {
   compile?: ng.IDirectiveCompileFn;
   link?: ng.IDirectiveLinkFn | ng.IDirectivePrePost;
+  [p: string]: any;
 }
 
 export function Directive({selector, ...options}: DirectiveOptionsDecorated) {
@@ -60,7 +62,42 @@ export function registerDirective(module: ng.IModule, ctrl: DirectiveControllerC
   }
   else {
     ctrl.$inject = ctrl.$inject || annotate(ctrl);
-    directiveFunc = () => ({...options, controller: ctrl});
+    const listeners: IHostListeners = getMetadata(metadataKeys.listeners, ctrl);
+    if (listeners) {
+      options.controller = extendWithHostListeners(ctrl, listeners);
+    }
+    directiveFunc = () => options;
   }
   module.directive(name, directiveFunc);
+}
+
+/** @internal */
+export function extendWithHostListeners(ctrl: {new(...args: any[])}, listeners: IHostListeners) {
+  const handlers = Object.keys(listeners);
+
+  class NewCtrl extends ctrl {
+    constructor(private $element, ...args: any[]) {
+      super(...args);
+    }
+    $postLink() {
+      if (super.$postLink) {
+        super.$postLink();
+      }
+      handlers.forEach(handler => {
+        const { eventName } = listeners[handler];
+        this.$element.on(eventName, this[handler].bind(this));
+      });
+    }
+    $onDestroy() {
+      if (super.$onDestroy) {
+        super.$onDestroy();
+      }
+      handlers.forEach(handler => {
+        const { eventName } = listeners[handler];
+        this.$element.off(eventName, this[handler]);
+      });
+    }
+  }
+  NewCtrl.$inject = ['$element', ...ctrl.$inject || []];
+  return NewCtrl;
 }