+++ src/module.ts
@@ -1,9 +1,10 @@
 import * as angular from 'angular';
 import { PipeTransform, registerPipe } from './pipe';
-import { ProviderObject, registerProviders } from './injectable';
+import { registerProviders } from './injectable';
 import { annotate, Declarations, getMetadata, metadataKeys } from './utils';
 import { registerComponent } from './component';
 import { registerDirective } from './directive';
+import { Provider } from './provider';
 
 export interface ModuleConfig {
   id?: string;
@@ -11,10 +12,10 @@ export interface ModuleConfig {
    * @deprecated
    */
   name?: string;
-  declarations: Array<ng.IComponentController | ng.Injectable<ng.IDirectiveFactory> | PipeTransform>;
+  declarations?: Array<ng.IComponentController | ng.Injectable<ng.IDirectiveFactory> | PipeTransform>;
   imports?: Array<string | NgModule>;
   exports?: Function[];
-  providers?: Array<ng.IServiceProvider | ng.Injectable<Function> | ProviderObject>;
+  providers?: Provider[];
 }
 
 export interface NgModule {
@@ -29,7 +30,7 @@ export interface NgModule {
   [s: string]: any;
 }
 
-export function NgModule({ id, name, declarations, imports = [], providers }: ModuleConfig) {
+export function NgModule({ id, name, declarations = [], imports = [], providers = [] }: ModuleConfig) {
   return (Class: NgModule) => {
     // module registration
     const deps = imports.map(mod => typeof mod === 'string' ? mod : mod.module.name);